requirements.txt
# ===========================================================================
# MDGT Edge Fingerprint Verification System - Python Dependencies
# ===========================================================================

# --- Web framework & server ------------------------------------------------
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0

# --- Async / networking ----------------------------------------------------
asyncssh>=2.14.0,<3.0.0
websockets>=12.0,<14.0
aiofiles>=23.2.0,<25.0

# --- Data validation -------------------------------------------------------
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# --- Cryptography ----------------------------------------------------------
cryptography>=41.0.0,<44.0

# --- Numerical / ML --------------------------------------------------------
numpy>=1.24.0,<2.1
pillow>=10.0.0,<12.0
faiss-cpu>=1.7.4,<2.0
onnxruntime>=1.16.0,<2.0

# --- CLI --------------------------------------------------------------------
click>=8.1.0,<9.0

# --- Configuration ----------------------------------------------------------
pyyaml>=6.0,<7.0

# --- Utilities (optional but recommended) -----------------------------------
httpx>=0.25.0,<1.0           # async HTTP client for health-checks / sync
python-multipart>=0.0.6      # file uploads via FastAPI
